Add SIGINT handler on posix systems
There's currently no overlap between the windows and posix
implementation of the exception handler so I split them into
different files for better clarity.

Added a signal handler for SIGINT to allow exiting on CTRL+C.
Note that like the rest of the program it does not exit cleanly and
triggers a SIGABRT.
